Implement Use More OOP and Organize into Packages
