+++ src/main/java/duke/Duke.java
@@ -1,3 +1,7 @@
+package duke;
+
+import duke.command.Command;
+
 import java.io.FileNotFoundException;
 
 public class Duke {
