+++ src/main/java/duke/command/ByeCommand.java
@@ -1,8 +1,13 @@
+package duke.command;
+
+import duke.Storage;
+import duke.TaskList;
+import duke.Ui;
+
 public class ByeCommand extends Command{
     public ByeCommand() {
         super(true);
     }
-
     @Override
     public void execute(TaskList tasks, Ui ui, Storage storage) {
         ui.showBye();
