+++ src/main/java/duke/command/EventCommand.java
@@ -1,10 +1,16 @@
+package duke.command;
+
 import java.time.LocalDate;
 
-public class DeadlineCommand extends Command{
+import duke.Storage;
+import duke.TaskList;
+import duke.Ui;
+
+public class EventCommand extends Command{
     private String description;
     private LocalDate date;
 
-    public DeadlineCommand(String description, LocalDate date) {
+    public EventCommand(String description, LocalDate date) {
         super(false);
         this.description = description;
         this.date = date;
@@ -12,7 +18,7 @@ public class DeadlineCommand extends Command{
 
     @Override
     public void execute(TaskList tasks, Ui ui, Storage storage) {
-        tasks.addDeadline(description, date);
+        tasks.addEvent(description, date);
         int size = tasks.getSize();
         ui.showAdd(tasks.getTask(size - 1), size);
         storage.save(tasks.saveToStorage());
