+++ src/main/java/duke/command/MarkCommand.java
@@ -1,7 +1,13 @@
-public class UnmarkCommand extends Command{
+package duke.command;
+
+import duke.DukeException;
+import duke.Storage;
+import duke.TaskList;
+import duke.Ui;
+public class MarkCommand extends Command{
     private int index;
 
-    public UnmarkCommand(int index) {
+    public MarkCommand(int index) {
         super(false);
         this.index = index;
     }
@@ -9,8 +15,8 @@ public class UnmarkCommand extends Command{
     @Override
     public void execute(TaskList tasks, Ui ui, Storage storage) throws DukeException {
         try {
-            tasks.unmarkTask(index);
-            ui.showUnmark(tasks.getTask(index));
+            tasks.markTask(index);
+            ui.showMark(tasks.getTask(index));
             storage.save(tasks.saveToStorage());
         } catch (IndexOutOfBoundsException e) {
             throw new DukeException("\tTask number does not exist!");
