+++ src/main/java/duke/command/ToDoCommand.java
@@ -1,3 +1,9 @@
+package duke.command;
+
+import duke.Storage;
+import duke.TaskList;
+import duke.Ui;
+
 public class ToDoCommand extends Command{
     private String description;
