+++ src/main/java/duke/command/UnmarkCommand.java
@@ -1,7 +1,14 @@
-public class MarkCommand extends Command{
+package duke.command;
+
+import duke.DukeException;
+import duke.Storage;
+import duke.TaskList;
+import duke.Ui;
+
+public class UnmarkCommand extends Command{
     private int index;
 
-    public MarkCommand(int index) {
+    public UnmarkCommand(int index) {
         super(false);
         this.index = index;
     }
@@ -9,8 +16,8 @@ public class MarkCommand extends Command{
     @Override
     public void execute(TaskList tasks, Ui ui, Storage storage) throws DukeException {
         try {
-            tasks.markTask(index);
-            ui.showMark(tasks.getTask(index));
+            tasks.unmarkTask(index);
+            ui.showUnmark(tasks.getTask(index));
             storage.save(tasks.saveToStorage());
         } catch (IndexOutOfBoundsException e) {
             throw new DukeException("\tTask number does not exist!");
