+++ src/main/java/duke/task/Deadline.java
@@ -1,5 +1,7 @@
-import java.time.format.DateTimeFormatter;
+package duke.task;
+
 import java.time.LocalDate;
+import java.time.format.DateTimeFormatter;
 
 public class Deadline extends Task {
     private LocalDate by;
